requirements.txt
# Core FastAPI dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# ASGI server and process management
gunicorn==21.2.0
python-dotenv==1.0.0

# Database and ORM
sqlalchemy==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.13.0  # Database migrations

# WebSocket and SSE support
python-socketio==5.11.0
websockets==12.0
sse-starlette==2.0.0

# Security - JWT and Password Hashing (UPDATED)
PyJWT==2.8.1
python-jose[cryptography]==3.3.0  # JWT with RSA support
passlib[argon2]==1.7.4  # Password hashing with Argon2
bcrypt==4.1.1  # Fallback password hashing
cryptography==41.0.7  # RSA key support
email-validator==2.1.0

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# Caching
redis==5.0.1
aiocache==0.12.2

# Observability and Monitoring
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-requests==0.42b0
prometheus-client==0.19.0

# Data processing
numpy==1.26.3
pandas==2.1.4

# GraphQL support
strawberry-graphql[fastapi]==0.219.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-benchmark==4.0.0
hypothesis==6.88.0  # Property-based testing

# Code quality and linting
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
ruff==0.1.8
bandit==1.7.5  # Security linting
safety==2.3.5  # Dependency vulnerability checking

# Pre-commit hooks
pre-commit==3.6.0

# Logging and debugging
python-json-logger==2.0.7
structlog==24.1.0

pylint==3.0.3

# Observability and monitoring
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
structlog==24.1.0

# Resilience patterns
tenacity==8.2.3
circuitbreaker==1.4.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Rate limiting
slowapi==0.1.9

# CORS
python-cors==1.0.0

# Environment variables
python-dotenv==1.0.0

# Validation
email-validator==2.1.0

# System monitoring
psutil==5.9.7

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
